feat: GPTQ-Int4 dequant + GEMV/GEMM kernels (Phase 1)
- gptq.h: GptqWeight descriptor, gptq_gemv/gemm/linear API, benchmark
- gptq.cu: CUDA kernels for SM87 (Orin)
  - GEMV: 64-col tile, 8-thread K-split, FP32 accum, shared mem reduction
  - GEMM: [32x64x128] tile, group-aligned BK=128, FP32 accum
  - Benchmark utility with FP64 CPU reference
- allocator.h: add INT32 DataType for packed GPTQ containers
- safetensors.cpp: map I32 dtype to INT32
- main.cpp: test-gptq (real weights) + bench-gptq commands
- Tegra: cudaHostRegister fails on read-only mmap; use cudaMemcpy
- Correctness: GEMV/GEMM PASS (max abs err 0.000038 vs FP64 ref)
- GEMV: ~52 GB/s (27% of 192 GB/s theory)
- GEMM: ~1.5 TFLOPS (29% of theory)

+++ src/machina/safetensors.cpp
@@ -187,7 +187,7 @@ DataType SafetensorsFile::parse_dtype(const std::string& s) const {
     if (s == "BF16") return DataType::BF16;
     if (s == "I8")   return DataType::INT8;
     if (s == "U8")   return DataType::U8;
-    if (s == "I32")  return DataType::FP32;  // treat as FP32 size for offsets
+    if (s == "I32")  return DataType::INT32;
     return DataType::UNKNOWN;
 }
